+++ nodes/Habitica/triggers/Common.ts
@@ -1,5 +1,8 @@
-import { IDataObject, INodePropertyOptions } from "n8n-workflow";
+import { IDataObject, IHookFunctions, INodePropertyOptions } from "n8n-workflow";
 import { triggerTaskScored, triggerTaskCreated, triggerTaskDeleted, triggerTaskUpdated } from './TriggerTaskActivity';
+import { triggerGroupChatMessageReceived } from "./TriggerGroupChat";
+import { triggerQuestFinished, triggerQuestInvited, triggerQuestStarted } from "./TriggerQuestActivity";
+import { triggerUserPetHatched, triggerUserMountRaised, triggerUserLeveledUp } from "./TriggerUserActivity";
 
 export interface IHabiticaTriggerBase {
 	/** n8n internal event name */
@@ -12,7 +15,7 @@ export interface IHabiticaTriggerBase {
 	triggerDefinition: INodePropertyOptions;
 
 	/** method for updating body parameters for "create webhook" request */
-	updateBodyParamsForCreateWebhook?(body: IDataObject): void;
+	updateBodyParamsForCreateWebhook?(body: IDataObject, hookFunctions: IHookFunctions): void;
 };
 
 // dictionary of all event definitions
@@ -22,4 +25,14 @@ export const allEventDefinitions: {[key: string]: IHabiticaTriggerBase} = {
 	[triggerTaskCreated.eventName]: triggerTaskCreated,
 	[triggerTaskUpdated.eventName]: triggerTaskUpdated,
 	[triggerTaskDeleted.eventName]: triggerTaskDeleted,
+	// group chat
+	[triggerGroupChatMessageReceived.eventName]: triggerGroupChatMessageReceived,
+	// quest activity
+	[triggerQuestStarted.eventName]: triggerQuestStarted,
+	[triggerQuestFinished.eventName]: triggerQuestFinished,
+	[triggerQuestInvited.eventName]: triggerQuestInvited,
+	// user activity
+	[triggerUserPetHatched.eventName]: triggerUserPetHatched,
+	[triggerUserMountRaised.eventName]: triggerUserMountRaised,
+	[triggerUserLeveledUp.eventName]: triggerUserLeveledUp,
 };