+++ CHANGELOG.md
@@ -1,1 +1,38 @@
 # Changelog
+
+## [0.2.0] - 2023-04-11
+
+All available triggers are supported.
+
+### Added
+
+* Added user activity triggers
+* Added quest activity triggers
+* Added group chat activity triggers
+
+## [0.1.0] - 2023-04-10
+
+### Added
+
+* Added task activity triggers
+
+## [0.0.2] - 2023-04-09
+
+### Tasks
+
+* Add a task
+* Get all tasks
+* Score a task
+
+### Skills (Spells)
+
+* Cast a skill
+
+### Chat messages
+
+* Send a chat message
+* Get all chat messages
+
+### Cron
+
+* Run cron